Create a skeleton for a new package that intends to use RcppArrayFire

RcppArrayFire.package.skeleton creates a skeleton for a new package that intends to use RcppArrayFire. It is based on the package.skeleton function that it executes first.

RcppArrayFire.package.skeleton(name = "anRpackage", list = character(),
  environment = .GlobalEnv, path = ".", force = FALSE,
  code_files = character(), example_code = TRUE)

Arguments

name
See package.skeleton

list
See package.skeleton

environment
See package.skeleton

path
See package.skeleton

force
See package.skeleton

code_files
See package.skeleton

example_code
If TRUE, example C++ code using RcppArrayFire is added to the package

Details

In addition to package.skeleton, the DESCRIPTION and NAMESPACE files are updated. In addition, minimal configure and src/Makevars.in files for finding the necessary compiler flags are created.